+++ src/store/modules/AudioManager.js
@@ -1,0 +1,13 @@
+import AudioManager from "@/classes/AudioManager";
+
+export default {
+    namespaced: true,
+    state: { audioManager: new AudioManager() },
+    getters: {
+        audioManager(state) {
+            return state.audioManager;
+        },
+    },
+    mutations: {},
+    actions: {},
+};